# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Portfolio API",
        "description": "Laravel12で構築したWeb APIのドキュメントです",
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "記事一覧を取得する",
                "operationId": "93ded9e31119e6bbb22455c72ef97f19",
                "parameters": [
                    {
                        "parameter": "ArticlePage",
                        "name": "page",
                        "in": "query",
                        "description": "ページ番号（デフォルト: 1）",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticlePagination"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "不正なパラメータ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "記事を投稿する",
                "operationId": "75bdc5f953261e6f34852386c9b522c0",
                "requestBody": {
                    "description": "記事情報",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostArticleDetailRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article created successfully."
                                        },
                                        "article": {
                                            "$ref": "#/components/schemas/ArticleDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "バリデーションエラー",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "指定した記事を取得する",
                "operationId": "a85697b2464e9aaa3f9980a83edbef88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "記事ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleDetailWithComments"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "不正なパラメータ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvalidParameter"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "記事が見つかりません",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "記事を更新する",
                "operationId": "beb39fc2f8d02c16b41438aadbd9ba03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "記事ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "記事情報",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostArticleDetailRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article created successfully."
                                        },
                                        "article": {
                                            "$ref": "#/components/schemas/ArticleDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "バリデーションエラー",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "記事を削除する",
                "operationId": "804276a4147e577744e158b2ce2b58f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "記事ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "記事が見つかりません",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleNotFound"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{id}/likes": {
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "記事にいいねを投稿する",
                "operationId": "d3669e60354903d8d729d021e9260329",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "記事ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article liked successfully."
                                        },
                                        "article_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "like": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "記事が見つかりません",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "PostArticleDetailRequest": {
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ArticleDetail": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "like": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ArticleDetailWithComments": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "記事タイトル"
                    },
                    "content": {
                        "type": "string",
                        "example": "記事の内容"
                    },
                    "like": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-23T07:54:58.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-23T07:54:58.000000Z"
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "article_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "message": {
                                    "type": "string",
                                    "example": "コメントの内容"
                                },
                                "created_at": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "updated_at": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "PaginationLink": {
                "properties": {
                    "url": {
                        "type": "string",
                        "nullable": true
                    },
                    "label": {
                        "type": "string"
                    },
                    "page": {
                        "type": "integer",
                        "nullable": true
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "ArticlePagination": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ArticleDetail"
                        }
                    },
                    "first_page_url": {
                        "type": "string"
                    },
                    "from": {
                        "type": "integer",
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "last_page_url": {
                        "type": "string"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PaginationLink"
                        }
                    },
                    "next_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "path": {
                        "type": "string",
                        "nullable": true
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 20
                    },
                    "prev_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "to": {
                        "type": "integer",
                        "example": 20
                    },
                    "total": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ArticleNotFound": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Article not found."
                    }
                },
                "type": "object"
            },
            "Unauthenticated": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "InvalidParameter": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Invalid parameter."
                    }
                },
                "type": "object"
            },
            "NotFound": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The title field is required. (and 1 more error)"
                    },
                    "errors": {
                        "description": "各フィールドごとのエラーメッセージ",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "The title field is required."
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctumで発行されたBearerトークンを入力してください",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}